[IMP] ODataRow: add copy constructor and methods
This patch adds a copy constructor to the ODataRow class. It will
allow to easily duplicate an ODataRow object, thanks to the _data of
the first one.

A "remove" method is also added to expose the epomynous method.

Finally the ODataRow internal _data HashMap is set as a private member
of the class.