[IMP] Use numeric keyboards for OInteger and OFloat types in a OField
Recent state: The OInteger and OFloat types in a OField have been mapped to a
OEditTextField that always used a normal text keyboard. Thus, it was possible
to enter non-numeric characters that cannot be stored in the database, so that
is was necessary to do input checks etc.

Now, new FieldTypes 'Integer' and 'Float' are introduced. In a OField, the
keyboard for a OInteger type is restricted to only digits, and the keyboard for
a OFloat type is restricted to digits and a (single) dot for fractional digits.