[IMP] Add getter for floats and avoid NullPointerExceptions in OValues
As OValues are closely related to ODataRecords which are using float values,
there should also be a getter for float values for OValues.

As all getters in OValues were using the wrapper classes for the primitive data
types, this commit also changes the return type for long values to 'Long'.

Until now, a NullPointerException has been thrown if someone tried to use the
typed getters in OValues for non-existant keys. This has been changed as well.
Instead, the methods are returning the same value as if the value of the field
would have been 'false'.